CODENVY-516: modify che-select widget for dashboard
