+++ dashboard/src/components/widget/select/che-select.html
@@ -1,12 +1,14 @@
 <div class="che-select">
   <!-- Mobile version -->
-  <md-input-container hide-gt-md class="md-input-has-value">
-    <label>{{labelName}}</label>
-    <select class="che-select-container"
-            ng-model="valueModel"
-            name="{{selectName}}">
-      <option value="">{{placeHolder}}</option>
-    </select>
+  <md-input-container class="che-select-mobile" hide-gt-md>
+    <label>{{value ? labelName : placeHolder}}</label>
+    <md-select ng-model="value"
+               md-container-class="che-custom-dropdown"
+               placeholder="{{placeHolder}}">
+      <md-option ng-value="optionValue.id ? optionValue.id : optionValue.name"
+                 ng-repeat="optionValue in optionValues">{{ optionValue.name }}
+      </md-option>
+    </md-select>
   </md-input-container>
 
   <!-- Desktop version -->
@@ -15,13 +17,13 @@
       <label flex="15">{{labelName}}: </label>
 
       <div layout="column" class="che-select-container" flex="85">
-        <select ng-model="valueModel"
-                name="{{selectName}}"
-                ng-mousedown="showOptions()"
-                ng-change="hideOptions()"
-                ng-blur="hideOptions()">
-          <option value="">{{placeHolder}}</option>
-        </select>
+        <md-select ng-model="value"
+                   md-container-class="che-custom-dropdown"
+                   placeholder="{{placeHolder}}">
+          <md-option ng-value="optionValue.id ? optionValue.id : optionValue.name"
+                     ng-repeat="optionValue in optionValues">{{ optionValue.name }}
+          </md-option>
+        </md-select>
       </div>
     </div>
   </div>
